yaml-cpp: fix build on newer CMake, add patch to supress debug messages
Upstream left these messages in, and they flood the build logs whenever
the library is used. Patch is taken from the upstream repo, however they
haven't done a release in a long time, and no release is in sight
either.

Also fixed build errors which are because of the old CMake requirement